Moved webpack to its own folder
