+++ package.json
@@ -7,10 +7,10 @@
     "postinstall": "bower install",
     "test": "karma start --single-run",
     "test-dev": "karma start",
-    "heroku-postbuild": "webpack -p --config webpack.config.prod.js",
+    "heroku-postbuild": "webpack -p --config webpack/config.prod.js",
     "start": "node server",
-    "build": "webpack --config webpack.config.dev.js",
-    "watch": "webpack --config webpack.config.dev.js --watch",
+    "build": "webpack --config webpack/config.dev.js",
+    "watch": "webpack --config webpack/config.dev.js --watch",
     "bower": "bower"
   },
   "author": "",
