+++ webpack/config.common.js
@@ -6,6 +6,7 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 const CleanWebpackPlugin = require("clean-webpack-plugin");
 const webpack = require("webpack");
 
+const projectRoot = path.join(__dirname, "../");
 const entryPoints = ["inline","polyfills","sw-register","styles","vendor","main"];
 const environmentName = process.env.NODE_ENV || "development";
 const isProdBuild = environmentName === "production";
@@ -15,19 +16,19 @@ console.log(environmentName);
 module.exports = {
     entry: {
         main: [
-            path.join(__dirname, "src/main.tsx")
+            path.join(projectRoot, "src/main.tsx")
         ],
         "service-worker": [
-            path.join(__dirname, "src/service-worker.ts")
+            path.join(projectRoot, "src/service-worker.ts")
         ]
     },
     output: {
-        path: path.join(__dirname, "dist"),
+        path: path.join(projectRoot, "dist"),
         filename: (details) => {
             if (details.chunk.name === "service-worker") {
                 return "[name].bundle.js";
             }
-            
+
             return `[name].bundle${isProdBuild ? ".[hash]" : ""}.js`;
         },
         chunkFilename: `[name].bundle${isProdBuild ? ".[hash]" : ""}.js`,
@@ -41,24 +42,24 @@ module.exports = {
             ".scss"
         ],
         modules: [
-            path.join(__dirname, "node_modules")
+            path.join(projectRoot, "node_modules")
         ],
         plugins: [
             new TsconfigPaths()
         ],
         alias: {
-            "@bootstrap": path.join(__dirname, "bower_components/bootstrap")
+            "@bootstrap": path.join(projectRoot, "bower_components/bootstrap")
         },
     },
     mode: environmentName,
     plugins: [
-        new CleanWebpackPlugin(path.join(__dirname, "dist")),
+        new CleanWebpackPlugin(path.join(projectRoot, "dist")),
         new webpack.NoEmitOnErrorsPlugin(),
         new MiniCssExtract({
             filename: `styles.bundle${isProdBuild ? ".[hash]" : ""}.css`
         }),
         new HtmlWebpackPlugin({
-            template: path.join(__dirname, "index.html"),
+            template: path.join(projectRoot, "src/index.html"),
             filename: "index.html",
             hash: false,
             inject: true,
